# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hệ Thống Tự Động Sinh Test Case API (CDIO-4)",
    "description": "Nền tảng tự động bóc tách đặc tả yêu cầu, giải ràng buộc và sinh bộ kiểm thử phần mềm đa phương pháp (BVA, Phân vùng tương đương, Pairwise, Z3 Solver). Quản lý chu kỳ Scrum, phân công nhiệm vụ thành viên và xuất báo cáo kiểm thử nghiệm thu chuyên nghiệp.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/export": {
      "get": {
        "tags": [
          "Quản lý Dự án"
        ],
        "summary": "Xuất báo cáo kiểm thử toàn diện cấp dự án",
        "description": "Xuất báo cáo nghiệm thu toàn bộ dự án dưới dạng Excel (4 sheet chuyên nghiệp), CSV, JSON, hoặc Markdown.",
        "operationId": "export_project_report_api_projects__project_id__export_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "xlsx",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export/{suite_id}": {
      "get": {
        "tags": [
          "Xuất Báo Cáo"
        ],
        "summary": "Xuất bộ ca kiểm thử (Excel 2 sheet, CSV, JSON, Markdown)",
        "description": "Xuất khẩu dữ liệu ca kiểm thử của bộ Test Suite ra các định dạng chuẩn: Excel chuyên nghiệp (Báo cáo tổng kết + Chi tiết ca test), CSV, JSON, hoặc bảng Markdown.",
        "operationId": "export_test_suite_api_export__suite_id__get",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suite Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "xlsx",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}